@@ -2,8 +2,6 @@
 
 namespace Armincms\EasyLicense\Listeners;
 
-use Illuminate\Contracts\Queue\ShouldQueue;
-use Illuminate\Queue\InteractsWithQueue;
 use Armincms\EasyLicense\Card;
 
 class OrderCompleted
@@ -15,36 +13,36 @@ class OrderCompleted
      * @return void
      */
     public function handle($event)
-    { 
-        if($event->order->orderable_type == \Armincms\EasyLicense\License::class) {
+    {
+        if ($event->order->orderable_type == \Armincms\EasyLicense\License::class) {
             $order = $event->order->loadMissing('saleables.saleable', 'customer');
 
-            $order->saleables->flatMap(function($orderItem) use ($order) {
-                if($orderItem->saleable->delivery === 'system') {
+            $order->saleables->flatMap(function ($orderItem) use ($order) {
+                if ($orderItem->saleable->delivery === 'system') {
                     return $orderItem->saleable->createCredit(
                         $order->customer, __('Online Payment'), $orderItem->count
-                    ); 
+                    );
                 }
-                if($orderItem->saleable->delivery === 'card') {
-                    $credits = Card::where('license_id', $orderItem->saleable->getKey())->whereHas('manuals', function($query) {
+                if ($orderItem->saleable->delivery === 'card') {
+                    $credits = Card::where('license_id', $orderItem->saleable->getKey())->whereHas('manuals', function ($query) {
                         return $query->forSales();
                     })->with([
-                        'manuals' => function($query) {
+                        'manuals' => function ($query) {
                             return $query->forSales();
-                        }
+                        },
                     ])->get()->flatMap->manuals;
-                    
-                    if ($credits->count() >= $orderItem->count && 
+
+                    if ($credits->count() >= $orderItem->count &&
                         collect($orderItem->details)->isEmpty()
-                    ) {  
+                    ) {
                         $orderItem->update([
-                            'details' => $credits->take($orderItem->count)->map->asSold()->each->save()->toArray()
+                            'details' => $credits->take($orderItem->count)->map->asSold()->each->save()->toArray(),
                         ]);
-                    } 
+                    }
                 }
-            })->filter()->each(function($credit) use ($order) {
+            })->filter()->each(function ($credit) use ($order) {
                 $credit->orders()->sync($order);
-            }); 
+            });
         }
     }
 }
